requirements.txt
Flask==2.3.2
flask-jwt-extended==4.7.1
flask-pymongo==2.3.0
Werkzeug==2.3.7
pymongo==4.4.0
python-dotenv==1.0.0
requests
